COMP: Remove signed to unsigned comparison warnings
Use explicit static_casts to compare signed to unsigned values.

+++ Libs/MRML/Core/vtkMRMLColorNode.cxx
@@ -783,14 +783,14 @@ void vtkMRMLColorNode::SetContainsTerminology(bool containsTerminology)
 //----------------------------------------------------------------------------
 void vtkMRMLColorNode::SetAllColorsDefined()
 {
-  int numberOfColors = this->GetNumberOfColors();
+  unsigned int numberOfColors = static_cast<unsigned int>(this->GetNumberOfColors());
   bool modified = false;
   if (this->Properties.size() < static_cast<size_t>(numberOfColors))
   {
     this->Properties.resize(numberOfColors);
     modified = true;
   }
-  for (int index = 0; index < numberOfColors; index++)
+  for (unsigned int index = 0; index < numberOfColors; index++)
   {
     if (this->Properties[index].Defined == true)
     {
